--- domain/doc.go
@@ -1,2 +1,0 @@
-// Package domain provides domain-specific types and functions.
-package domain
