+++ examples/veterinarian/main.go
@@ -8,7 +8,6 @@ import (
 	"os"
 
 	semantic_router "github.com/conneroisu/go-semantic-router"
-	"github.com/conneroisu/go-semantic-router/domain"
 	"github.com/conneroisu/go-semantic-router/encoders/ollama"
 	"github.com/conneroisu/go-semantic-router/stores/memory"
 	"github.com/ollama/ollama/api"
@@ -18,7 +17,7 @@ import (
 // noteworthy here means that the routes are likely to be relevant to a noteworthy conversation in a veterinarian appointment.
 var NoteworthyRoutes = semantic_router.Route{
 	Name: "noteworthy",
-	Utterances: []domain.Utterance{
+	Utterances: []semantic_router.Utterance{
 		{Utterance: "what is the best way to treat a dog with a cold?"},
 		{Utterance: "my cat has been limping, what should I do?"},
 	},
@@ -28,7 +27,7 @@ var NoteworthyRoutes = semantic_router.Route{
 // chitchat here means that the routes are likely to be relevant to a chitchat conversation in a veterinarian appointment.
 var ChitchatRoutes = semantic_router.Route{
 	Name: "chitchat",
-	Utterances: []domain.Utterance{
+	Utterances: []semantic_router.Utterance{
 		{Utterance: "what is your favorite color?"},
 		{Utterance: "what is your favorite animal?"},
 	},
