+++ route_test.go
@@ -4,7 +4,6 @@ import (
 	"testing"
 
 	semanticrouter "github.com/conneroisu/go-semantic-router"
-	"github.com/conneroisu/go-semantic-router/domain"
 	"github.com/conneroisu/go-semantic-router/encoders/ollama"
 	"github.com/conneroisu/go-semantic-router/stores/memory"
 	"github.com/ollama/ollama/api"
@@ -15,7 +14,7 @@ import (
 // noteworthy here means that the routes are likely to be relevant to a noteworthy conversation in a veterinarian appointment.
 var NoteworthyRoutes = semanticrouter.Route{
 	Name: "noteworthy",
-	Utterances: []domain.Utterance{
+	Utterances: []semanticrouter.Utterance{
 		{Utterance: "what is the best way to treat a dog with a cold?"},
 		{Utterance: "my cat has been limping, what should I do?"},
 	},
@@ -25,7 +24,7 @@ var NoteworthyRoutes = semanticrouter.Route{
 // chitchat here means that the routes are likely to be relevant to a chitchat conversation in a veterinarian appointment.
 var ChitchatRoutes = semanticrouter.Route{
 	Name: "chitchat",
-	Utterances: []domain.Utterance{
+	Utterances: []semanticrouter.Utterance{
 		{Utterance: "what is your favorite color?"},
 		{Utterance: "what is your favorite animal?"},
 	},
