--- transformers.go
@@ -1,8 +1,0 @@
-package semanticrouter
-
-import "context"
-
-// Chatter is an interface that defines a method, Chat, which takes a string
-type Chatter interface {
-	Chat(ctx context.Context, utterance string) (string, error)
-}
